Create access rule

+++ resources/views/post/show.blade.php
@@ -1,14 +1,16 @@
 @extends('app')
 
 @section('content')
-    <button type="button" class="btn btn-primary" onclick="window.location='{{ URL::route('posts.edit',$post->id) }}'">
-        Update
-    </button>
-    {!! Form::open(array('route' => array('posts.destroy', $post->id), 'method' => 'delete')) !!}
-    <button type="submit" class="btn btn-primary">
-        Delete
-    </button>
-    {!! Form::close() !!}
+    @if(Entrust::hasRole('admin'))
+        <button type="button" class="btn btn-primary" onclick="window.location='{{ URL::route('posts.edit',$post->id) }}'">
+            Update
+        </button>
+        {!! Form::open(array('route' => array('posts.destroy', $post->id), 'method' => 'delete')) !!}
+        <button type="submit" class="btn btn-primary">
+            Delete
+        </button>
+        {!! Form::close() !!}
+    @endif
     <article>
         <h2>
             {{ $post->title }}
